Fix file paths in handler.js When moving files around we neglected to go up another directory after nesting the handler.js file another directory deep, thus our serveLanding and servePublic functions did not work. This commit fixes this. fixes #71
