+++ src/server.js
@@ -1,5 +1,5 @@
 const http = require('http');
-const router = require('./router.js');
+const router = require('./routers/router.js');
 
 const host = process.env.HOST || 'localhost';
 const port = process.env.PORT || 4000;
